Fix to allow setting IDs for newly created records
Previously, saving a new record would trigger a POST request to
the FHIR server. In the FHIR specification, the only way to create
a new record and assign it an ID is by using a PUT request. This
fix causes the adapter to use a PUT request when saving a new
record that has an ID that has been manually set. If no ID is set,
it defaults to a POST request which will auto-generate an ID.